requirements.txt
python-binance>=1.0.19
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
